requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
blinker==1.6.1
Brotli==1.0.9
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
configparser==5.3.0
crayons==0.4.0
cryptography==40.0.1
exceptiongroup==1.1.1
fake-headers==1.0.2
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
multidict==6.0.4
oauthlib==3.2.2
outcome==1.2.0
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.27.1
requests-oauthlib==1.3.1
selenium==4.7.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
trio==0.22.0
trio-websocket==0.10.2
twitter-scraper-selenium==4.1.4
typing_extensions==4.5.0
urllib3==1.26.15
webdriver-manager==3.2.2
webencodings==0.5.1
wsproto==1.2.0
yarl==1.8.2
zstandard==0.20.0
